refactor(desktop): extract stream forwarding into its own module
commands.rs had grown to 1560 lines, well past the 800-line threshold in
AGENTS.md, and the stream forwarding loops had no module of their own to
carry their tests or invariants.

Move forward_contract_stream, forward_workspace_watch, and to_contract_change
into stream_forwarding.rs unchanged. The repeated registry cleanup and the
spelled-out registry type collapse into an unregister helper and a
StreamRegistry alias; behavior is identical.
